Add and Subtract Logic Added
Calculator now can add and subtract simple numbers. It cannot perform the
operation multiple times and add the same numbers over and over. If the numbers are too big the calculator will
not recognize them as a valid entry. Need to implement logic for
division and multiplication.

Adding a result button would be benefitial to optimize the calculator.